Rename `-utils` to `utils`
